refactor: move independent logic into separate modules
